% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{: fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of a state in the US as given in the FARS data
sets. (This represents the state that the user wishes to plot)}

\item{year}{The year of analysis}
}
\value{
Returns a map plot of all the traffic accidents that happened in the
given state and year. Returns an error "invalid STATE number" if the state
number isn't in the FARS data set. Returns "no accidents to plot" if there
are no accidents in the state for that given year (This might happen if
the user selects a year that is outside of the FARS dataset
}
\description{
This function plots, on a map, all the accidents that happened
in a given US state for the given year.
}
\examples{
\dontrun{
fars_map_state(2,2014)
}

}
