% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string with file name of the CSV file to be read.}
}
\value{
Returns a data.frame based on the CSV file.
If there is not file, a "does not exist" error is read
}
\description{
{fars_read} Loads a CSV file specifically data from the Fatality Analysis Reporting System
run by the US National Highway Traffice Safety Administration. If the file path is incorrect
the function returns a "does not exist" error.
}
\examples{
\dontrun{
fars_read("accident_2014.csv.bz2")
}


}
