% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{: fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years in numeric format.}
}
\value{
If the file exists, returns data frame with MONTH and year of
each of the entries in that data. Otherwise, returns NULL (i.e. error)
}
\description{
The function accepts a vector or list of years and returns a list of data
frames with MONTH and year columns based on data in "accident_<year>.csv.bz2
files. The files need to be located in the working directory.

Function takes in a list of years, and returns a list of data frames with
two columns, MONTH and year. '
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
