% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{: fars_summarise_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years (i.e. the function will then show the number
of accidents in each month for the given years)}
}
\value{
Data frame with months in rows and years in columns. Each value shows
the number oaf ccidents in that given month-year.
}
\description{
Tabulates the number of accidents in the month-year combination
e.g. how many accidents happened in Apr 2014.'
}
\examples{
\dontrun{
fars_summarize_years(2012:2014)
}

}
