% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Number representing the year of the accidents in the data set.}
}
\value{
Returns a file name based on the year provided.
#'
}
\description{
Creates a file name for the uploaded CSV file.
This is based on the year of the accidents, documented in the CSV file
}
\examples{
\dontrun{
makefilename(2014)
}

}
